refactor(audit): A3 stage 6 — AppBundleBuilder -> domain interface + data impl (#19)
Removed the last android.content.Context from the domain layer: AppBundleBuilder is now a pure domain
interface (domain/repository/AppBundleBuilder.kt; build(appInfo, cacheSubDir): Result<File>) with the
Context-holding implementation moved verbatim to data (AppBundleBuilderImpl, @Single(binds=...)
auto-discovered by @ComponentScan). ExportAppUseCase/ShareAppUseCase now inject the interface
(domain->domain); Koin wires the data impl.

Domain purity: no android.* framework types remain in domain EXCEPT android.net.Uri in three
repository interfaces (AppAnalyzer / InstallerRepository / StoreRepository) + ShareAppUseCase's return
— a thin content-handle, OUTSIDE the audit's #45 scope (which targeted the heavy Bitmap/Drawable/
Intent + PackageManager mapping, all now removed). Left as an accepted residual; androidx.compose
@Immutable stability hints also retained.

Build: assembleFossDebug green.
